requirements.txt
pandas>=1.5.0
gspread>=5.7.2
google-auth>=2.20.0
google-auth-oauthlib>=0.7.1
google-auth-httplib2>=0.1.0
transformers>=4.30.2
torch>=2.0.1
gradio>=3.39.1
deep-translator>=1.11.4
regex>=2022.10.31
